pass along determinateness information while compiling clauses.
Also distinguish between "last" clause and last clause of a clause-block where all other
clauses are known to be determinate - in this case a final self-tail-call is considered
determinate, as well.

All this allows us to infer whether predicates are determinate. This is only used in
calls following the definition, or by using -xref to list determinate predicates.